back-end/requirements.txt
Django
djangorestframework
psycopg2-binary
django-probes
httpie
django-cors-headers
